v0.3.2
legit why do i care about such niche bugs that genuinely don't matter. It works now so ig that's that, but was it worth FOUR HOURS of my sanity? no. it was not. a totally unfair exchange in the realm of alchemy if you ask me. AND it build for the last push?? i have it- uk what. I'm not a dev ops engineer. it's not my bare minimum to only build when /docs is pushed to; that's the expected. alright whatever rant over, thank you for coming to my ted talk